Added insert_after method
The first overload for the insert after method takes  a valid iterator in the range [begin, end) and returns a new list with an element added right after the position pointed by it.

Some changes were made to immutable_list. It is now a friend of immutable_list_iterator to access its private node member to more easily and efficiently work with iterator positions.
A new constructor was added to create a new list from a valid iterator range.
Finally, a tail member was added for O(1) cend support.

immutable_list_iterator::operator++(int) was corrected. It previously, wrongly, returned a reference instead of a copy of the previous position of the iterator, resulting in a memory error as the referenced instance would be destroyed at the end of the call.